renaming files to match coding standards
