+++ client/helpers/get-string-elipsis.js
