+++ client/helpers/index.js
@@ -1,4 +1,4 @@
-export { default as getJsonStringObject } from './getJsonStringObject';
-export { default as getKeyValuePairs } from './getKeyValuePairs';
-export { default as getStringElipsis } from './getStringElipsis';
-export { default as mapDomainDescription } from './mapDomainDescription';
+export { default as getJsonStringObject } from './get-json-string-object';
+export { default as getKeyValuePairs } from './get-key-value-pairs';
+export { default as getStringElipsis } from './get-string-elipsis';
+export { default as mapDomainDescription } from './map-domain-description';
