+++ client/helpers/map-domain-description.js
