+++ client/routes/execution/helpers/get-event-details.js
