+++ client/routes/execution/helpers/get-time-elapsed-display.js
