+++ client/routes/execution/helpers/get-time-stamp-display.js
