+++ client/routes/execution/helpers/index.js
@@ -1,13 +1,13 @@
-export { default as getEventDetails } from './getEventDetails';
-export { default as getEventFullDetails } from './getEventFullDetails';
-export { default as getEventSummary } from './getEventSummary';
-export { default as getHistoryEvents } from './getHistoryEvents';
-export { default as getHistoryTimelineEvents } from './getHistoryTimelineEvents';
-export { default as getSummary } from './getSummary';
-export { default as getSummaryWorkflowStatus } from './getSummaryWorkflowStatus';
-export { default as getTimeElapsedDisplay } from './getTimeElapsedDisplay';
-export { default as getTimeStampDisplay } from './getTimeStampDisplay';
-export { eventFullTransforms } from './eventFullTransforms';
-export { default as mapTimelineEvents } from './mapTimelineEvents';
-export { default as parentWorkflowLink } from './parentWorkflowLink';
-export { summarizeEvents } from './summarizeEvents';
+export { default as getEventDetails } from './get-event-details';
+export { default as getEventFullDetails } from './get-event-full-details';
+export { default as getEventSummary } from './get-event-summary';
+export { default as getHistoryEvents } from './get-history-events';
+export { default as getHistoryTimelineEvents } from './get-history-timeline-events';
+export { default as getSummary } from './get-summary';
+export { default as getSummaryWorkflowStatus } from './get-summary-workflow-status';
+export { default as getTimeElapsedDisplay } from './get-time-elapsed-display';
+export { default as getTimeStampDisplay } from './get-time-stamp-display';
+export { eventFullTransforms } from './event-full-transforms';
+export { default as mapTimelineEvents } from './map-timeline-events';
+export { default as parentWorkflowLink } from './parent-workflow-link';
+export { summarizeEvents } from './summarize-events';
